feat: addition of historical graph and layout improvement

+++ src/Potentiometer.c
@@ -8,8 +8,8 @@ void configure_potentiometer()
 }
 
 // Função para ler a posição do potenciômetro e retornar o valor
-float read_potentiometer() {
-    float sum = 0.0f;
+uint read_potentiometer() {
+    uint sum = 0.0f;
     adc_select_input(2);
 
     for (int i = 0; i < 100; i++)
@@ -18,18 +18,18 @@ float read_potentiometer() {
         vTaskDelay(pdMS_TO_TICKS(10));
     }
 
-    float average = sum / 100.0f;
+    uint average = sum / 100.0f;
     
-    printf("adc: %.2f\n", average);
+    printf("adc: %d\n", average);
 
     return average;
 }
 
 // Função para mapear o valor lido do potenciômetro para um intervalo específico
-float map_reading(float value, float in_min, float in_max, float out_min, float out_max) {
+uint map_reading(uint value, uint in_min, uint in_max, uint out_min, uint out_max) {
     // Verifica se o valor está dentro do intervalo de entrada
     if (value < in_min || value > in_max) {
-        return -1.0f; // Retorna -1 se o valor estiver fora do intervalo
+        return -1; // Retorna -1 se o valor estiver fora do intervalo
     }
     
     // Mapeia o valor do intervalo de entrada para o intervalo de saída
